+++ internal/features/profiles/repository/profile.go
@@ -22,9 +22,9 @@ func NewProfileRepository(postgres *sql.DB, redis *redis.Client) *profileReposit
 
 type IProfileRepository interface {
 	FindAllProfiles(ctx context.Context) ([]models.Profile, error)
-	FindProfileById(ctx context.Context, profileId uint) (models.Profile, error)
+	FindProfileById(ctx context.Context, profileId int) (models.Profile, error)
 	FindProfileByEmail(ctx context.Context, email string) (models.Profile, error)
-	CreateProfile(ctx context.Context, profile *models.Profile) (int64, error)
-	UpdateProfileById(ctx context.Context, profileId uint, profileRequest *models.ProfileUpdateRequest) (models.Profile, error)
-	DeleteProfileById(ctx context.Context, profileId uint) (bool, error)
+	CreateProfile(ctx context.Context, profile *models.Profile) (int, error)
+	UpdateProfileById(ctx context.Context, profileId int, profileRequest *models.ProfileUpdateRequest) (models.Profile, error)
+	DeleteProfileById(ctx context.Context, profileId int) (bool, error)
 }
