+++ internal/shared/tokens/tokens.go
@@ -18,9 +18,9 @@ const (
 )
 
 type ITokens interface {
-	SaveRefreshToken(ctx context.Context, profileId int64, refreshToken string) error
-	DeleteRefreshToken(ctx context.Context, profileId int64, refreshToken string) error
-	RetrieveRefreshToken(ctx context.Context, profileId int64, refreshToken string) (string, error)
+	SaveRefreshToken(ctx context.Context, profileId int, refreshToken string) error
+	DeleteRefreshToken(ctx context.Context, profileId int, refreshToken string) error
+	RetrieveRefreshToken(ctx context.Context, profileId int, refreshToken string) (string, error)
 }
 
 type tokens struct {
@@ -33,7 +33,7 @@ func NewTokens(redis *redis.Client) *tokens {
 	}
 }
 
-func (t *tokens) SaveRefreshToken(ctx context.Context, profileId int64, refreshToken string) error {
+func (t *tokens) SaveRefreshToken(ctx context.Context, profileId int, refreshToken string) error {
 	err := t.redis.HSet(ctx, fmt.Sprintf("%d", profileId), "refreshToken", refreshToken)
 
 	if err != nil {
@@ -42,7 +42,7 @@ func (t *tokens) SaveRefreshToken(ctx context.Context, profileId int64, refreshT
 	return nil
 }
 
-func (t *tokens) RetrieveRefreshToken(ctx context.Context, profileId int64, refreshToken string) (string, error) {
+func (t *tokens) RetrieveRefreshToken(ctx context.Context, profileId int, refreshToken string) (string, error) {
 	rt, err := t.redis.Get(ctx, constants.REFRESH_TOKEN_KEY).Result()
 	if err != redis.Nil {
 		return "", fmt.Errorf("%s key does not exist: %w", constants.REFRESH_TOKEN_KEY, err)
@@ -52,6 +52,6 @@ func (t *tokens) RetrieveRefreshToken(ctx context.Context, profileId int64, refr
 	return rt, nil
 }
 
-func (t *tokens) DeleteRefreshToken(ctx context.Context, profileId int64, refreshToken string) error {
+func (t *tokens) DeleteRefreshToken(ctx context.Context, profileId int, refreshToken string) error {
 	return nil
 }
