+++ internal/shared/types.go
@@ -1,5 +1,5 @@
 package shared
 
 type ItemID struct {
-	ID int64 `json:"item_id"`
+	ID int `json:"item_id"`
 }
